+++ src/pages/CompareJson/utils/getIndent.js
@@ -1,0 +1,6 @@
+export const getIndent = (level) => {
+  if (level === 1) {
+    return { textIndent: "20px" };
+  }
+  return { textIndent: `${level * 20}px` };
+};
