+++ src/pages/CompareJson/utils/index.js
@@ -1,0 +1,5 @@
+export * from "./getIndent";
+export * from "./getType";
+export * from "./isArray";
+export * from "./isObject";
+export * from "./needFormat";
