+++ src/pages/CompareJson/utils/isArray.js
@@ -1,0 +1,6 @@
+export const isArray = (item) => {
+  if (item === "array") {
+    return true;
+  }
+  return Object.prototype.toString.call(item) === "[object Array]";
+};
