+++ src/pages/CompareJson/utils/isObject.js
@@ -1,0 +1,3 @@
+export const isObject = (item) => {
+  return Object.prototype.toString.call(item) === "[object Object]";
+};
