+++ src/pages/CompareJson/utils/needFormat.js
@@ -1,0 +1,3 @@
+export const needFormat = (type) => {
+  return type === "array" || type === "object";
+};
